backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pillow==10.2.0
pillow-heif==0.14.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.3
